Add 3 tests in cell_spec and Update README
Hey team; I wrote 3 tests based on the Interaction Pattern and added some notes to the README. Not sure If I'll keep updating the README, but hopefully it makes things a little easier. Communication is important, and I think this could get complicated, depending on how far we get. Thoughts?

[Ticket: X]